requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
psycopg2-binary
python-jose
passlib
python-multipart
pytest-mock
pytest-asyncio